+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.3.0] - 2026-09-03
+
 ### Added
 
 - `mpvd pick` Ctrl+a keybind to toggle all filtered files.
@@ -131,7 +133,8 @@ and this project adheres to
 - `remove`/`rm` subcommand
 - `send` subcommand
 
-[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.2...main
+[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.3.0...main
+[2.3.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.2...v2.3.0
 [2.2.2]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.1...v2.2.2
 [2.2.1]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.1.1...v2.2.0